requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
numpy>=1.26.0
pillow>=10.0.0
tqdm>=4.66.0
matplotlib>=3.8.0
scipy>=1.11.0
imageio>=2.34.0
gradio>=4.44.0
